+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to semantic versioning.
 
 - Add `until` to `Monad`, `IO`, and `Task`.
 - Add `singleton` and an `Eq` instance to `URLSearchParams`.
+- Rename `getParam` to `lookupFirst` in `URLSearchParams`.
 
 ## 0.17.1 (2023-06-17)